Added more consistent handling of slide shows
(i.e. individual video frames to be shown until the next video frame
appears in the stream). This allows correct playback of still frames
with or without audio regardless of domain (previously still frames
with audio was only handled in a menu), so this change allows audio
DVDs to be played.

from:
https://github.com/MythTV/mythtv/commit/4df1612e567c1c5e2f19fc312ffef4532fb61bae

references:
https://code.mythtv.org/trac/ticket/11579